Assignment2/requirements.txt
numpy>=1.24
opencv-contrib-python>=4.8
scikit-learn>=1.3
matplotlib>=3.7
PyYAML>=6.0
